Multiple changes (see full commit message)
The signature for `metadata` has been changed to only require `Self`
instead of `Self::Metadata`. This permits `Metadata` to capture the
lifetime of the type it is for, rather than taking the user's word on
what it should be. The final change here is that `new` now takes `self`
as a parameter for the same reason. It is not stored in memory.

There are three bug fixes present. One is that the width of integer is
now correct. Another is that the flags for `FormatterOptions` were not
properly obtained in the getters. The final fix is the bounds of
`Metadata` being relaxed to what is strictly necessary.

Finally, this commit implements a method to obtain the padded width of a
value.